Move postgresqlSecret function to postgresql.go
